Fix : 스케쥴 추가, 수정, 디테일, 버그 수정

+++ app/src/main/java/com/sas/companymanagement/ui/artist/db/ArtistDao.kt
@@ -30,5 +30,5 @@ interface ArtistDao {
     fun getAllArtist(): LiveData<List<Artist>>
 
     @Query("SELECT * FROM artist_tbl WHERE id = :id")
-    fun findArtistById(id: Int): LiveData<Artist>
+    fun findArtistById(id: Long): LiveData<Artist>
 }
