+++ app/src/main/java/com/sas/companymanagement/ui/artist/update/ArtistUpdateViewModel.kt
@@ -30,7 +30,7 @@ class ArtistUpdateViewModel(application: Application) : AndroidViewModel(applica
     fun updateArtist(artist: Artist) {
         repository.updateUser(artist)
     }
-    fun findArtistById(id: Int): LiveData<Artist> {
+    fun findArtistById(id: Long): LiveData<Artist> {
         return repository.findArtistById(id)
     }
 }
